packages/shared/src/utils/schedules.test.ts
import moment from "moment";
import { ScheduleRecurrence } from "../api";
import { calculateNextDailyOccurrence, calculateNextMonthlyOccurrence, calculateNextWeeklyOccurrence, calculateNextYearlyOccurrence, validateTimeFrame } from "./schedules";

describe("validateTimeFrame", () => {
    beforeAll(() => {
        // eslint-disable-next-line @typescript-eslint/no-empty-function
        jest.spyOn(console, "error").mockImplementation(() => { });
    });
    afterAll(() => {
        jest.restoreAllMocks();
    });

    it("should return true for valid timeframes within a year", () => {
        const start = new Date("2022-01-01T00:00:00Z");
        const end = new Date("2022-12-31T00:00:00Z");
        expect(validateTimeFrame(start, end)).toBeTruthy();
    });

    it("should return false for timeframes longer than a year", () => {
        const start = new Date("2022-01-01T00:00:00Z");
        const end = new Date("2023-01-02T00:00:00Z");
        expect(validateTimeFrame(start, end)).toBeFalsy();
    });

    it("should return true for exact one-year timeframes", () => {
        const start = new Date("2022-01-01T00:00:00Z");
        const end = new Date("2023-01-01T00:00:00Z");
        expect(validateTimeFrame(start, end)).toBeTruthy();
    });

    it("should return true for very short timeframes", () => {
        const start = new Date("2022-01-01T00:00:00Z");
        const end = new Date("2022-01-01T00:01:00Z"); // 1 minute later
        expect(validateTimeFrame(start, end)).toBeTruthy();
    });

    it("should return false for timeframes with start date after end date", () => {
        const start = new Date("2022-01-02T00:00:00Z");
        const end = new Date("2022-01-01T00:00:00Z"); // Start is after end
        expect(validateTimeFrame(start, end)).toBeFalsy();
    });
});

type CalculateOccurrenceFn = (currentDate: Date, recurrence: ScheduleRecurrence, timeZone?: string) => Date;
const testRecurrence = ({
    currentDate,
    expected,
    func,
    recurrence,
    timeZone,
}: {
    currentDate: string,
    expected: string,
    func: CalculateOccurrenceFn,
    recurrence: ScheduleRecurrence,
    timeZone?: string,
}): void => {
    const current = new Date(currentDate);
    const result = func(current, recurrence, timeZone);
    const expectedResult = moment.tz(expected, timeZone ?? "UTC").toDate();
    console.log("result", result);
    expect(result.toISOString()).toBe(expectedResult.toISOString());
};

describe("calculateNextDailyOccurrence", () => {
    it("should return the next occurrence for a 1-day interval", () => {
        testRecurrence({
            currentDate: "2023-01-01T00:02:00Z",
            expected: "2023-01-02T00:02:00Z",
            func: calculateNextDailyOccurrence,
            recurrence: {
                recurrenceType: "Daily",
                interval: 1,
            } as ScheduleRecurrence,
            timeZone: "Poland",
        });
    });

    it("should return the next occurrence for a 3-day interval", () => {
        testRecurrence({
            currentDate: "2023-03-10T00:00:00Z",
            expected: "2023-03-13T00:00:00Z",
            func: calculateNextDailyOccurrence,
            recurrence: {
                recurrenceType: "Daily",
                interval: 3,
            } as ScheduleRecurrence,
            timeZone: "America/New_York",
        });
    });

    it("should handle leap years correctly", () => {
        testRecurrence({
            currentDate: "2024-02-28T00:30:02Z",
            expected: "2024-02-29T00:30:02Z",
            func: calculateNextDailyOccurrence,
            recurrence: {
                recurrenceType: "Daily",
                interval: 1,
            } as ScheduleRecurrence,
            timeZone: "US/Mountain",
        });
    });

    it("should handle end of month correctly", () => {
        testRecurrence({
            currentDate: "2023-04-30T00:00:00Z",
            expected: "2023-05-01T00:00:00Z",
            func: calculateNextDailyOccurrence,
            recurrence: {
                recurrenceType: "Daily",
                interval: 1,
            } as ScheduleRecurrence,
            timeZone: "Indian/Christmas",
        });
    });
});

describe("calculateNextWeeklyOccurrence", () => {

    test("next occurrence without specified dayOfWeek", () => {
        testRecurrence({
            currentDate: "2023-10-31T12:01:01Z",
            expected: "2023-11-07T12:01:01Z",
            func: calculateNextWeeklyOccurrence,
            recurrence: {
                recurrenceType: "Weekly",
                interval: 1,
            } as ScheduleRecurrence,
            timeZone: "Africa/Nairobi",
        });
    });

    test("next occurrence with specified dayOfWeek that is in the future of the current week", () => {
        testRecurrence({
            currentDate: "2023-10-31T12:00:00Z", // Tuesday
            expected: "2023-11-02T12:00:00Z",
            func: calculateNextWeeklyOccurrence,
            recurrence: {
                recurrenceType: "Weekly",
                interval: 1,
                dayOfWeek: 4, // Thursday
            } as ScheduleRecurrence,
            timeZone: "America/Mexico_City",
        });
    });

    test("next occurrence with specified dayOfWeek that has already passed in the current week", () => {
        testRecurrence({
            currentDate: "2023-10-31T12:00:00Z", // Tuesday
            expected: "2023-11-06T12:00:00Z",
            func: calculateNextWeeklyOccurrence,
            recurrence: {
                recurrenceType: "Weekly",
                interval: 1,
                dayOfWeek: 1, // Monday
            } as ScheduleRecurrence,
            timeZone: "America/New_York",
        });
    });

    test("next occurrence with dayOfWeek same as current day", () => {
        testRecurrence({
            currentDate: "2023-10-31T12:00:00Z", // Tuesday
            expected: "2023-11-07T12:00:00Z",
            func: calculateNextWeeklyOccurrence,
            recurrence: {
                recurrenceType: "Weekly",
                interval: 1,
                dayOfWeek: 2, // Tuesday
            } as ScheduleRecurrence,
            // timeZone: "Pacific/Auckland", TODO breaks - off by one day
        });
    });

    test("next occurrence with interval greater than 1 and no dayOfWeek", () => {
        testRecurrence({
            currentDate: "2023-10-31T12:00:00Z",
            expected: "2023-11-21T12:00:00Z",
            func: calculateNextWeeklyOccurrence,
            recurrence: {
                recurrenceType: "Weekly",
                interval: 3,
            } as ScheduleRecurrence,
            timeZone: "America/Argentina/San_Juan",
        });
    });

    test("next occurrence with interval greater than 1 and dayOfWeek", () => {
        testRecurrence({
            currentDate: "2023-10-31T12:00:00Z", // Tuesday
            expected: "2023-11-02T12:00:00Z",
            func: calculateNextWeeklyOccurrence,
            recurrence: {
                recurrenceType: "Weekly",
                interval: 2,
                dayOfWeek: 4, // Thursday
            } as ScheduleRecurrence,
            timeZone: "Asia/Rangoon",
        });
    });

});

describe("calculateNextMonthlyOccurrence", () => {
    test("should calculate the next occurrence in the current month, when the day hasn't happened yet", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:00:00Z",
            expected: "2023-01-20T00:00:00Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 1,
                dayOfMonth: 20,
            } as ScheduleRecurrence,
        });
    });

    test("should calculate the next occurrence in the next month, when the day has already passed", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:08:09Z",
            expected: "2023-02-10T00:08:09Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 1,
                dayOfMonth: 10,
            } as ScheduleRecurrence,
            // timeZone: "America/Mexico_City", //TODO breaks - off by one day
        });
    });

    test("should calculate the next occurrence for the same day when dayOfMonth is not set", () => {
        testRecurrence({
            currentDate: "2023-01-15T01:00:00Z",
            expected: "2023-02-15T01:00:00Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 1,
            } as ScheduleRecurrence,
            timeZone: "Africa/Nairobi",
        });
    });

    test("should handle the case where the dayOfMonth exceeds the number of days in the next month", () => {
        testRecurrence({
            currentDate: "2023-02-15T08:32:00Z",
            expected: "2023-03-03T08:32:00Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 1,
                dayOfMonth: 31,
            } as ScheduleRecurrence,
            timeZone: "America/New_York",
        });
    });

    test("should find the event later this month, even with the interval of 2", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:16:00Z",
            expected: "2023-01-20T00:16:00Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 2,
                dayOfMonth: 20,
            } as ScheduleRecurrence,
            timeZone: "Pacific/Auckland",
        });
    });

    test("should find the event 2 months from now, since the dayOfMonth has passed", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:00:00Z",
            expected: "2023-03-10T00:00:00Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 2,
                dayOfMonth: 10,
            } as ScheduleRecurrence,
            timeZone: "Europe/Oslo",
        });
    });

    test("should calculate the next occurrence 4 months later when dayOfMonth is not set", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:00:07Z",
            expected: "2023-05-15T00:00:07Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 4,
            } as ScheduleRecurrence,
            timeZone: "America/Mexico_City",
        });
    });

    test("should calculate the next occurrence 4 months later with dayOfMonth set", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:00:07Z",
            expected: "2023-05-14T00:00:07Z",
            func: calculateNextMonthlyOccurrence,
            recurrence: {
                recurrenceType: "Monthly",
                interval: 4,
                dayOfMonth: 14,
            } as ScheduleRecurrence,
            timeZone: "Europe/London",
        });
    });
});

describe("calculateNextYearlyOccurrence", () => {
    test("should calculate the next occurrence for a specific month and day of the month", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:00:00Z",
            expected: "2023-06-20T00:00:00Z",
            func: calculateNextYearlyOccurrence,
            recurrence: {
                recurrenceType: "Yearly",
                interval: 1,
                month: 5, // June
                dayOfMonth: 20,
            } as ScheduleRecurrence,
            timeZone: "Europe/London",
        });
    });

    test("should calculate the next occurrence for the same day and month when month and dayOfMonth are not set", () => {
        testRecurrence({
            currentDate: "2023-01-15T00:05:00Z",
            expected: "2024-01-15T00:05:00Z",
            func: calculateNextYearlyOccurrence,
            recurrence: {
                recurrenceType: "Yearly",
                interval: 1,
            } as ScheduleRecurrence,
            timeZone: "Australia/North",
        });
    });

    test("should handle the case where the specified date is already passed in the current year", () => {
        testRecurrence({
            currentDate: "2023-07-15T00:00:10Z",
            expected: "2024-06-20T00:00:10Z",
            func: calculateNextYearlyOccurrence,
            recurrence: {
                recurrenceType: "Yearly",
                interval: 1,
                month: 5, // June
                dayOfMonth: 20,
            } as ScheduleRecurrence,
            // timeZone: "Canada/Saskatchewan", TODO breaks - off by one day
        });
    });
});
